Move node module file to root
